[FIX] refactor entity type definitions for AttributionAdd components to improve clarity and maintainability
fixes:
/frontend/src/components/Attribution/Add/AttributionAddFormWrapper.vue
  92:7  error  'AttributionAddFormProps' is an 'error' type that acts as 'any' and overrides all other types in this union type  @typescript-eslint/no-redundant-type-constituents
